move ios-scroll-freeze-bug
