+++ test/functional/test-ios-scrollfreeze-bug.js
@@ -15,7 +15,7 @@
  */
 
 import {AmpDocSingle} from '../../src/service/ampdoc-impl';
-import {checkAndFix} from '../../src/service/ios-scrollfreeze-bug';
+import {checkAndFix} from '../../src/service/viewport/ios-scrollfreeze-bug';
 import * as sinon from 'sinon';
